+++ lib/src/isolate/hand_detector_core.dart
@@ -92,7 +92,7 @@ class HandDetectorCore {
     required double gestureMinConfidence,
     bool useCompiledModel = false,
     Set<Accelerator> accelerators = const {Accelerator.gpu, Accelerator.cpu},
-    Precision precision = Precision.fp16,
+    Precision precision = Precision.fp32,
   }) async {
     _mode = mode;
     _maxDetections = maxDetections;
